+++ asus-nb/src/dbus_ledmode.rs
@@ -6,15 +6,25 @@ use dbus::blocking;
 
 pub trait OrgAsuslinuxDaemon {
     fn set_led_mode(&self, data: &str) -> Result<(), dbus::Error>;
+    fn next_led_mode(&self) -> Result<(), dbus::Error>;
+    fn prev_led_mode(&self) -> Result<(), dbus::Error>;
     fn led_mode(&self) -> Result<String, dbus::Error>;
     fn led_modes(&self) -> Result<String, dbus::Error>;
-    fn led_bright(&self) -> Result<i16, dbus::Error>;
+    fn led_brightness(&self) -> Result<i16, dbus::Error>;
 }
 
 impl<'a, T: blocking::BlockingSender, C: ::std::ops::Deref<Target=T>> OrgAsuslinuxDaemon for blocking::Proxy<'a, C> {
 
     fn set_led_mode(&self, data: &str) -> Result<(), dbus::Error> {
         self.method_call("org.asuslinux.Daemon", "SetLedMode", (data, ))
+    }
+
+    fn next_led_mode(&self) -> Result<(), dbus::Error> {
+        self.method_call("org.asuslinux.Daemon", "NextLedMode", ())
+    }
+
+    fn prev_led_mode(&self) -> Result<(), dbus::Error> {
+        self.method_call("org.asuslinux.Daemon", "PrevLedMode", ())
     }
 
     fn led_mode(&self) -> Result<String, dbus::Error> {
@@ -27,8 +37,8 @@ impl<'a, T: blocking::BlockingSender, C: ::std::ops::Deref<Target=T>> OrgAsuslin
             .and_then(|r: (String, )| Ok(r.0, ))
     }
 
-    fn led_bright(&self) -> Result<i16, dbus::Error> {
-        self.method_call("org.asuslinux.Daemon", "LedBright", ())
+    fn led_brightness(&self) -> Result<i16, dbus::Error> {
+        self.method_call("org.asuslinux.Daemon", "LedBrightness", ())
             .and_then(|r: (i16, )| Ok(r.0, ))
     }
 }